解决或运算告警

+++ services/src/data_center/form_info/form_info_helper.cpp
@@ -83,7 +83,7 @@ ErrCode FormInfoHelper::LoadFormConfigInfoByBundleNames(const std::vector<std::s
     }
 
     std::vector<BundleInfo> bundleInfos;
-    int32_t flag = GET_BUNDLE_INFO_WITH_ALL_EXTENSIONS | GET_BUNDLE_INFO_EXCLUDE_EXT;
+    uint32_t flag = (uint32_t)GET_BUNDLE_INFO_WITH_ALL_EXTENSIONS | (uint32_t)GET_BUNDLE_INFO_EXCLUDE_EXT;
     ErrCode ret = IN_PROCESS_CALL(iBundleMgr->BatchGetBundleInfo(bundleNames, flag,
         bundleInfos, userId));
     HILOG_INFO("bundleInfos size:%{public}zu, bundleNames size:%{public}zu", bundleInfos.size(), bundleNames.size());
